Added method TokenizeStringSingleQuotes
Added a new method TokenizeStringSingleQuotes()
to handle strings with single quotes.

Now strings are defined with double quotes and single
quotes. Each can contain the other but not itself.
For example a double quote embedded string can not contain
a double quote.

"This is a string with a '"' double quote". This would lead
to an error.