Changed function
Changed the tokenize() function.

The actual tokenization is now done by the GetNextToken()
method. This method gets the current token and matches it
against the defined rules.

This way the tokenize function is much shorter.

Added token kinds _EOF, UNDERSCORE.

The _EOF token is always appended at the end to signalize the end of
the token stream for the AST creation.

The underscore must be handled separately now, because the GetNextToken()
methods handles the character matching differently.